Extract coefficients of a DiscSurv object

Return the model coefficients of a DiscSurv object

Usage

# S3 method for class 'DiscSurv'
coef(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)

Arguments

fit

a DiscSurv object.

lambda

values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.

which

indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.

drop

whether to keep coefficient names

...

Examples

data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
fit <- DiscSurv(data, Event.char, Z.char, Time.char)
coef(fit, lambda = fit$lambda)$alpha[, 1:10]
#>                 0.4261    0.3882    0.3537    0.3223    0.2937    0.2676
#> [Time: 0.53] -1.979501 -1.970055 -1.977003 -1.996580 -2.025985 -2.063312
#> [Time: 1.03] -2.110213 -2.069229 -2.043936 -2.030857 -2.027099 -2.024915
#> [Time: 3.92] -2.484907 -2.418504 -2.367851 -2.329504 -2.300521 -2.271900
#> [Time: 6.74] -2.355695 -2.272311 -2.204520 -2.148985 -2.102801 -2.057118
#> [Time: 12.5] -2.958691 -2.863140 -2.784495 -2.719193 -2.664049 -2.607962
#>                 0.2438    0.2221    0.2024    0.1844
#> [Time: 0.53] -2.108360 -2.159506 -2.215515 -2.275258
#> [Time: 1.03] -2.030033 -2.041091 -2.057155 -2.077143
#> [Time: 3.92] -2.250843 -2.236081 -2.226730 -2.221891
#> [Time: 6.74] -2.019057 -1.987374 -1.961171 -1.939712
#> [Time: 12.5] -2.560104 -2.519177 -2.484277 -2.454635
coef(fit, lambda = fit$lambda)$beta[, 1:10]
#>    0.4261     0.3882     0.3537     0.3223       0.2937      0.2676     0.2438
#> Z1      0 -0.1556065 -0.2988513 -0.4321738 -0.555170528 -0.62756301 -0.6972625
#> Z2      0  0.0000000  0.0000000  0.0000000  0.000000000  0.00000000  0.0000000
#> Z3      0  0.0000000  0.0000000  0.0000000  0.000000000  0.00000000  0.0000000
#> Z4      0  0.0000000  0.0000000  0.0000000  0.000000000  0.00000000  0.0000000
#> Z5      0  0.0000000  0.0000000  0.0000000 -0.003092387 -0.06677843 -0.1269445
#>        0.2221     0.2024       0.1844
#> Z1 -0.7646826 -0.8306135 -0.891613065
#> Z2  0.0000000  0.0000000  0.000000000
#> Z3  0.0000000  0.0000000 -0.005933632
#> Z4  0.0000000  0.0000000  0.000000000
#> Z5 -0.1842370 -0.2383347 -0.287641261