Extract coefficients of a ppDiscSurv object

Return the model coefficients of a ppDiscSurv object

Usage

# S3 method for class 'ppDiscSurv'
coef(fit, lambda, which = 1:length(fit$lambda), drop = TRUE, ...)

Arguments

fit

a ppDiscSurv object.

lambda

values of the regularization parameter lambda at which coefficients are requested. For values of lambda not in the sequence of fitted models, linear interpolation is used.

which

indices of the penalty parameter lambda at which predictions are required. By default, all indices are returned. If lambda is specified, this will override which.

drop

whether to keep coefficient names

...

Examples

data(DiscTime)
data <- DiscTime$data
Event.char <- DiscTime$Event.char
prov.char <- DiscTime$prov.char
Z.char <- DiscTime$Z.char
Time.char <- DiscTime$Time.char
fit <- pp.DiscSurv(data, Event.char, prov.char, Z.char, Time.char)
coef(fit, lambda = fit$lambda)$alpha[, 1:5]
#>                 0.1601    0.1458    0.1329    0.1211    0.1103
#> [Time: 0.53] -1.668065 -1.747753 -1.824209 -1.898663 -1.971909
#> [Time: 1.03] -1.480994 -1.526815 -1.571328 -1.615488 -1.659886
#> [Time: 3.92] -1.622145 -1.641556 -1.661848 -1.683563 -1.706991
#> [Time: 6.74] -1.251585 -1.257340 -1.264124 -1.272503 -1.282795
#> [Time: 12.5] -1.781635 -1.780095 -1.780526 -1.783340 -1.788771
coef(fit, lambda = fit$lambda)$gamma[, 1:5]
#>       0.1601     0.1458    0.1329     0.1211     0.1103
#> 1  0.0000000  0.0000000  0.000000  0.0000000  0.0000000
#> 2 -4.5665147 -4.3516946 -4.162992 -3.9945779 -3.8425705
#> 3 -0.7478311 -0.7116142 -0.682481 -0.6580980 -0.6367834
#> 4  1.2412090  1.1456446  1.059286  0.9815198  0.9121089
#> 5 -2.3399410 -2.1966372 -2.070536 -1.9577270 -1.8555876
coef(fit, lambda = fit$lambda)$beta[, 1:5]
#>    0.1601     0.1458     0.1329     0.1211     0.1103
#> Z1      0 -0.1735757 -0.3325774 -0.4793679 -0.6155647
#> Z2      0  0.0000000  0.0000000  0.0000000  0.0000000
#> Z3      0  0.0000000  0.0000000  0.0000000  0.0000000
#> Z4      0  0.0000000  0.0000000  0.0000000  0.0000000
#> Z5      0  0.0000000  0.0000000  0.0000000  0.0000000